99-openroast.rules: Add udev rules for the sr700
Running on my Debian system I'm running OpenRoast as a non-root user. As
a result I don't have permissions to read and write the ttyUSB device.
In order to fix this a udev rule is required.

This patch adds a udev rule for the SR700. I'm not sure myself what the
other VID/PIDs for the other supported SRXXX devices are but, it is
trivial to add those in by others who have this data.

Fix for the sr700 now.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@nexus-software.ie>